# server/openapi.yaml
openapi: 3.1.0
info:
  title: EV2 Clandestinoz API
  version: 0.1.0
  description: |
    API del sistema de gestión de nightclub EV2 Clandestinoz.

    **Este archivo es el contrato entre backend, web y apps móviles.** Cualquier cambio
    de ruta o de forma de respuesta se refleja aquí en el mismo commit que lo introduce.

    ## Autenticación
    Salvo las rutas marcadas como públicas, toda petición requiere
    `Authorization: Bearer <access_token>`. El token de acceso dura 15 minutos; se renueva
    con `POST /api/auth/refresh` usando el refresh token, que es de **un solo uso**
    (al canjearlo se revoca y se emite uno nuevo).

    ## Errores
    Todos los errores tienen la misma forma (`ErrorResponse`). Los códigos usados son
    400 validación, 401 sin autenticar, 403 sin permiso, 404 no existe, 409 conflicto de
    estado, 422 regla de negocio, 429 límite de peticiones, 500 interno, 501 no implementado.

    ## Dinero
    Los importes son cadenas decimales con dos decimales (`"1500.00"`) para no perder
    precisión, y siempre van acompañados de `currency` (`MXN` o `USD`). Los precios se
    calculan **siempre en el servidor**.

    ## Precios de reservación
    Una mesa se reserva por **toda la noche**, que corresponde a un evento del calendario.
    El precio es `base de la zona + (personas extra x precio del boleto de esa noche) +
    extras`. Cada zona incluye un número de boletos y admite un tope de extras (por
...

...

    ## Flirt
...

    ## Empleados y dinero
...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/nightclubs/{nightclubId}:
    get:
      operationId: getNightclubsByNightclubId
      tags: [Club]
      summary: Perfil del club
      parameters: [{ $ref: '#/components/parameters/NightclubId' }]
      responses:
        '200':
          description: Perfil
          content:
            application/json:
              schema:
                type: object
                properties:
                  nightclub: { $ref: '#/components/schemas/Nightclub' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/nightclubs/{nightclubId}/orders:
    post:
      operationId: postNightclubsByNightclubIdOrders
      tags: [Pedidos]
      summary: Crear un pedido de bebidas
      description: |
        Idempotente por `client_request_id`. Verifica existencias con bloqueo de fila
        (dos pedidos simultáneos no pueden agotar el mismo producto), congela el precio
        de cada artículo y descuenta el inventario en la misma transacción.

        Si `recipient_id` está presente, el pedido es una invitación a otra persona; se
        rechaza si esa persona bloqueó al remitente.

        **El pedido nace con su cobro.** Junto con el pedido se crea un asiento en
        `transactions` (`reference_type: drink_order`) en estado `pending`, y la barra no
        puede confirmarlo mientras no esté `paid`. El cobro se registra en la mesa con
        `POST /manual-payments/register` (efectivo o terminal del club), y pagarlo
        confirma el pedido solo: nadie tiene que apretar un segundo botón.

        **El mesero levanta el pedido de la mesa que atiende** enviando `table_id` y,
        cuando el cliente está registrado y sentado ahí, `on_behalf_of`. Sin
        `on_behalf_of` el cobro queda a nombre del mesero, que es quien recibió el
        dinero. `taken_by` en la respuesta dice quién lo levantó.
      parameters: [{ $ref: '#/components/parameters/NightclubId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrderCreate' }
      responses:
        '201':
          description: Pedido creado
          content:
            application/json:
              schema:
                type: object
                properties:
                  order: { $ref: '#/components/schemas/Order' }
        '200':
          description: |
            Reenvío de una petición ya procesada. Devuelve el pedido original con la
            cabecera `Idempotent-Replay: true`.
          headers:
            Idempotent-Replay:
              schema: { type: string, enum: ['true'] }
          content:
            application/json:
              schema:
                type: object
                properties:
                  order: { $ref: '#/components/schemas/Order' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Producto no disponible o sin existencias
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example:
                error:
                  code: conflict
                  message: Some drinks are not available
                  details:
                    - drink_id: 0a1b2c3d-0000-4000-8000-000000000001
                      name: Cerveza nacional
                      reason: out_of_stock
                      stock: 0
        '422': { $ref: '#/components/responses/Unprocessable' }
    get:
      operationId: getNightclubsByNightclubIdOrders
      tags: [Pedidos]
      summary: Cola de pedidos
      description: Requiere rol `bartender`, `waiter` o `manager`.
      parameters:
        - { $ref: '#/components/parameters/NightclubId' }
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, preparing, ready, delivered, cancelled, pos_error]
        - name: active
          in: query
          description: Solo pedidos sin terminar
          schema: { type: boolean, default: false }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
      responses:
        '200':
          description: Pedidos, del más antiguo al más reciente
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items: { $ref: '#/components/schemas/Order' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/nightclubs/{nightclubId}/orders/mine:
    get:
      operationId: getNightclubsByNightclubIdOrdersMine
      tags: [Pedidos]
      summary: Mis pedidos (enviados y recibidos)
      parameters:
        - { $ref: '#/components/parameters/NightclubId' }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
      responses:
        '200':
          description: Pedidos
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items: { $ref: '#/components/schemas/Order' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/nightclubs/{nightclubId}/orders/{orderId}:
    get:
      operationId: getNightclubsByNightclubIdOrdersByOrderId
      tags: [Pedidos]
      summary: Detalle de un pedido
      description: Accesible para quien lo envió, quien lo recibe y el personal.
      parameters:
        - { $ref: '#/components/parameters/NightclubId' }
        - { $ref: '#/components/parameters/OrderId' }
      responses:
        '200':
          description: Pedido
          content:
            application/json:
              schema:
                type: object
                properties:
                  order: { $ref: '#/components/schemas/Order' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/nightclubs/{nightclubId}/orders/{orderId}/status:
    post:
      operationId: postNightclubsByNightclubIdOrdersByOrderIdStatus
      tags: [Pedidos]
      summary: Cambiar el estado de un pedido
      description: |
        Transiciones permitidas:

        | Desde | Hacia |
        |---|---|
        | `pending` | `confirmed`, `cancelled` |
        | `confirmed` | `preparing`, `cancelled` |
        | `preparing` | `ready`, `cancelled` |
        | `ready` | `delivered` |
        | `pos_error` | `confirmed`, `cancelled` |

        El personal (`bartender`, `waiter`, `manager`) puede hacer cualquiera de ellas.
        Un cliente solo puede **cancelar su propio pedido mientras siga en `pending`**.
        Cancelar devuelve las existencias al inventario.
      parameters:
        - { $ref: '#/components/parameters/NightclubId' }
        - { $ref: '#/components/parameters/OrderId' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [confirmed, preparing, ready, delivered, cancelled]
                reason: { type: string, maxLength: 200 }
      responses:
        '200':
          description: Estado actualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  order:
                    type: object
                    properties:
                      id: { type: string, format: uuid }
                      status: { type: string }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token de acceso de 15 minutos obtenido en login o refresh.
    agentKey:
      type: apiKey
      in: header
      name: X-Agent-Key
      description: |
        Llave propia del agente local del POS. Autentica una maquina, no una persona:
        no hereda permisos de ningun usuario y solo sirve en las rutas /api/pos/agent/*.
  parameters:
    NightclubId:
      name: nightclubId
      in: path
      required: true
      description: Debe coincidir con el club del usuario autenticado; si no, 403.
      schema: { type: string, format: uuid }
...
    OrderId:
      name: orderId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
...
  responses:
    BadRequest:
      description: Validación fallida
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            error:
              code: bad_request
              message: Validation failed
              details:
                - field: body.email
                  message: must be a valid email
    Unauthorized:
      description: Falta el token o no es válido
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Sin permiso (rol insuficiente u otro club)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: El recurso no existe
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Conflicto con el estado actual del recurso
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unprocessable:
      description: Contraviene una regla de negocio
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [bad_request, unauthorized, forbidden, not_found, conflict,
                     unprocessable, too_many_requests, internal_error, not_implemented]
            message: { type: string }
            details:
              description: Presente en errores de validación y de disponibilidad.
              type: array
              items: { type: object, additionalProperties: true }
            request_id:
              type: string
              description: Identificador de la petición, útil para buscar en los registros.
    Money:
      type: string
      pattern: '^\d+\.\d{2}$'
      description: Importe decimal con dos decimales, como cadena para no perder precisión.
      example: '1500.00'
    Currency:
      type: string
      enum: [MXN, USD]
...
    NightclubPublic:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        city: { type: string }
        country: { type: string }
        timezone: { type: string, example: America/Hermosillo }
        currency_default: { $ref: '#/components/schemas/Currency' }
    Nightclub:
      allOf:
        - { $ref: '#/components/schemas/NightclubPublic' }
        - type: object
          properties:
            address: { type: string }
            phone: { type: string }
            capacity: { type: integer }
            settings: { type: object, additionalProperties: true }
...
    OrderItemInput:
      type: object
      required: [drink_id]
      properties:
        drink_id: { type: string, format: uuid }
        quantity: { type: integer, minimum: 1, maximum: 20, default: 1 }
        notes: { type: string, maxLength: 120 }
    OrderCreate:
      type: object
      required: [client_request_id, items]
      properties:
        client_request_id:
          type: string
          format: uuid
          description: UUID generado por el cliente; garantiza que no se dupliquen pedidos.
        table_id: { type: string, format: uuid }
        recipient_id:
          type: string
          format: uuid
          description: Para invitar una bebida a otra persona.
        on_behalf_of:
          type: string
          format: uuid
          description: >
            Solo personal. El cliente sentado en `table_id` para quien se levanta el
            pedido: el cobro queda a su nombre. Si la persona no está en esa mesa: 422.
        message: { type: string, maxLength: 280 }
        items:
          type: array
          minItems: 1
          maxItems: 20
          items: { $ref: '#/components/schemas/OrderItemInput' }
    Order:
      type: object
      properties:
        id: { type: string, format: uuid }
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, delivered, cancelled, pos_error]
        subtotal: { $ref: '#/components/schemas/Money' }
        currency: { $ref: '#/components/schemas/Currency' }
        message: { type: [string, 'null'] }
        table_id: { type: [string, 'null'], format: uuid }
        table_code: { type: [string, 'null'] }
        sender_id: { type: string, format: uuid }
        sender_name: { type: [string, 'null'] }
        recipient_id: { type: [string, 'null'], format: uuid }
        recipient_name: { type: [string, 'null'] }
        bartender_id: { type: [string, 'null'], format: uuid }
        taken_by:
          type: [string, 'null']
          format: uuid
          description: Mesero que levantó el pedido. Nulo cuando el cliente lo pidió desde su propio teléfono.
        taken_by_name: { type: [string, 'null'] }
        transaction_id:
          type: [string, 'null']
          format: uuid
          description: El cobro del pedido. Nulo solo si el pedido no cuesta nada.
        payment_status:
          type: string
          enum: [not_required, pending, pending_manual, paid, failed, refunded, cancelled]
          description: >
            Estado del cobro. La barra no puede confirmar el pedido mientras no sea `paid`;
            pagarlo lo confirma solo. `not_required` es un pedido sin costo.
        returned_to_sender:
          type: boolean
          description: Invitación rechazada por el receptor y devuelta a la mesa del emisor (nunca cancelada ni reembolsada).
        returned_at: { type: [string, 'null'], format: date-time }
        pos_order_id:
          type: [string, 'null']
          description: Identificador en el punto de venta (fase 4).
        pos_error: { type: [string, 'null'] }
        items:
          type: array
          items:
            type: object
            properties:
              drink_id: { type: string, format: uuid }
              name: { type: string }
              quantity: { type: integer }
              unit_price: { $ref: '#/components/schemas/Money' }
              notes: { type: [string, 'null'] }
        created_at: { type: string, format: date-time }
        confirmed_at: { type: [string, 'null'], format: date-time }
        ready_at: { type: [string, 'null'], format: date-time }
        delivered_at: { type: [string, 'null'], format: date-time }
        cancelled_at: { type: [string, 'null'], format: date-time }
...